requirements.txt
asgiref==3.2.10
backcall==0.2.0
decorator==4.4.2
Django==3.1.2
djangorestframework==3.12.1
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
jedi==0.18.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
psycopg2==2.8.6
ptyprocess==0.7.0
Pygments==2.8.1
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
